msm: camera: isp: Fix no free payload issue
Currently there is flush workqueue operation during stop. When flush
if the workqueue is queued and not to be running, flush operation
will cancel the workqueue. Due to get the csid path payload at the
top half and put the payload at the bottom half, so if the workqueue
is canceled, there is not chance to put the payload. For the long
time, the payload will be exhausted and then no free payload issue
will happen. This commit reinitialize the free payload list at the
end of stop. This commit also add get_async_task for no crm state
machine and crm flush to let flush before stop to avoid workqueue
execute after reinitialize the free payload list.

CRs-Fixed: 3526854
Change-Id: I561e709d75b182e6680fa76c86aa1a7f6069a444
Signed-off-by: zhuo <quic_zhuo@quicinc.com>