msm: camera: icp: Reorganize ICP V2 driver for Matrix
Add support to handle a variation of ICP V2 for matrix.
IRQ registers have been moved to a different reg space.
Some registers that are currently used for debug  will
not be accessible by HLOS on matrix, remove usage of them.
The change also removes dependency on the reg base order
populated in the DTSI, irrespective of the order the driver
will parse and find the respective base indexes.
The change also adds register specific info per chipset.

CRs-Fixed: 3628787
Change-Id: I59545dac77954eeee95b5057134682d867e10988
Signed-off-by: Stuti Saxena <quic_stutsaxe@quicinc.com>