msm: camera: reqmgr: Fix frame drop issue
When the req id of a slot is -1, we don't update the
last applied idx for this slot, if a bubble issue is
reported on previous slot, then we won't reset its
status to ADDED, then we will force reset it during
moving to new slot, then this slot will be reset, its
status will be NO_REQ after reset, then wr idx will be
increased even though the wr idx isn't equal to rd idx.
This change updates the wr idx only when wr idx is equal
to rd idx, and always update last applied idx.

CRs-Fixed: 3788288
Change-Id: Ia0946647e1629ab2bfa1e3eb68ddadd9510a98b0
Signed-off-by: zhuo <quic_zhuo@quicinc.com>