qcx_kernel: Added cci suspend function to only enter suspend when clocks are disabled as part of CPMS suspend(userspace suspend).
Description: When suspend is triggered, userspace sends a cmd to kernel to turn off all the necessary clocks. There are cases where turning off the clocks will not be completed but device enters suspend, this results in suspend failure.

Solution: Added a cci suspend function, which checks for the state of the cci device and only enter suspend when the clocks are off
Change-Id: I397268fd8d299747511223fcc5b98e579ad113a5
Signed-off-by: Moin Mustak Goveri <quic_mgoveri@quicinc.com>