msm: camera: cci: Semaphore may sleep in spin lock
When we enable CONFIG_DEBUG_ATOMIC_SLEEP=y, the kernel will do
might_sleep check when we down a semaphore, if the might_sleep
found it is in atomic context, it will eventually call BUG() to
raise kernel panic because sleep is not permitted in atomic
context. What we do now is get a spin lock before down a
semaphore, this is possible to make might_sleep raise a
kernel panic because semaphore might sleep in spin lock. So now
change the spin lock to mutex lock to avoid this situation.

CRs-Fixed: 3867582
Signed-off-by: Jaydeep Bagthariya <quic_jbagthar@quicinc.com>
Change-Id: If558e5f4916a17ba8757440158ff15fd0995f4df