msm: camera: isp: Error Handling for Sensor Foveation
If userspace does not send a packet on frame N, there
is no setting ID buffer received on N+1 frame and
we don't have a sensor Setting ID to compare while
applying request.

To fix this issue, this commit maintains a scratch
buffer, programs it for sensor metadata resource
and uses it for such case.

CRs-Fixed: 3871023
Change-Id: I976fcbdcafec134bc33f57c3acb0b6385ef4fade
Signed-off-by: Stuti Saxena <quic_stutsaxe@quicinc.com>