msm: camera: isp: Correct PHY mask to support more CSIPHY Hardware
Currently full ife/csid and IFE/CSID lite uses differnte header files
for memory mapping and other operational prospectives. IFE and CSID
lite is only enabled for the 3 CSIPHY hardware, whereas there are 5
CSIPHY hardware is avaialable. This causes the acquire failure when
CSIPHY index 3 or 4 is being requested. This change fixes this issue
for both Full and Lite IFE at the time of reservging the CID resource
based on Titan and CPAS hw version.

Change-Id: I8cddf138794cbede96e7eb4fcb51be28b78a06d0
Signed-off-by: Changxiang Wang <quic_changxia@quicinc.com>