msm: camera: isp: Fix RUP BL preparation in UL path
For UL path, RUP command is prepared once in a session
and reused continuously. Currently, this BL command was
been created in first setting KMD buffer, which is getting
overwritten later due to recycling of KMD buffer. This
is resulting in invalid command error from CDM.

Moving the RUP command BL to INIT KMD buffer, which is
not recycled during session.

CRs-Fixed: 3974898
Change-Id: Ie00d78fe54afb8e50a64ba1d8d2862f9bd15394d
Signed-off-by: Rishabh Jain <quic_risjai@quicinc.com>