msm: camera: common: Fix slab-out-of-bounds in cam_packet validate
Fix slab-out-of-bounds in cam_packet_util_validate. This issue was due to
KMD copying the packet from UMD but the length of packet we got was from
UMD which could have been corrupted, and then we were reading this copied
packet but the lengths here were based on buffer length, due to which a
out of bound might have occurred when copied length was very small value
compared to cam_packet. This change ensures that the packet length value
we get from UMD is  greater than or equal to size of cam_packet struct.

CRs-Fixed: 3953618
Change-Id: I712fc20914a4f5cc242482235793a12a9589181d
Signed-off-by: Ashish Bhimanpalliwar <quic_abhiman@quicinc.com>
Signed-off-by: Pranav Sanwal <quic_psanwal@quicinc.com>
Signed-off-by: Shivakumar Malke <quic_smalke@quicinc.com>
Signed-off-by: Zun Qiao <quic_zqiao@quicinc.com>
(cherry picked from commit 5e90453cdae4b78fd6f1fb5e77932bcb29f2d2fb)