msm: camera: common: Fix slab-out-of-bounds in cam_packet validate
Fix slab-out-of-bounds in cam_packet_util_validate. This issue was due to
KMD copying the packet from UMD but the length of packet we got was from
UMD which could have been corrupted, and then we were reading this copied
packet but the lengths here were based on buffer length, due to which a
out of bound might have occurred when copied length was very small value
compared to cam_packet. This change ensures that the packet length value
we get from UMD is  greater than or equal to size of cam_packet struct.

CRs-Fixed: 3953618

Change-Id: If213c9cfb85fc9984b6b83ccc9c0e95e77ac6767
Signed-off-by: Ashish Bhimanpalliwar <quic_abhiman@quicinc.com>
Signed-off-by: Pranav Sanwal <quic_psanwal@quicinc.com>
Signed-off-by: Shivakumar Malke <quic_smalke@quicinc.com>
Signed-off-by: Pooja Chitakoti <quic_pchitako@quicinc.com>