msm: camera: isp: Early return from no_crm apply in case of EPOCH clubbing
In case of bottom half scheduling delay, bottom half for epoch
are getting clubbed. Due to which, in UL path the pending list
was getting appended with new requests but they were not applied to HW.
This was resulting into improper updates for producer queue, which was
resulting into frame drops.

This change avoids appending new request to pending list in case of
epoch clubbing.

CRs-Fixed: 3999544
Change-Id: Ic3e9f06c9f4bedf5c1e1b06c4000fa4ace3f0ed3
Signed-off-by: Rishabh Jain <quic_risjai@quicinc.com>