msm: camera: isp: Add SFE Lite support for offline read
The SFE LITE fetch is mainly used for offline image reads.
This change enables offline reading, which is handled by
the SFE_LITE read engine and then fed to the IFE for
processing. This change activates the offline ISP feature.
We have also refactored the code based on requirements,
as the SFE_LITE hardware only includes a read engine.

CRs-Fixed: 3987317
Change-Id: I1cdd1bec3654bf629a90c362b850286da58b9fee
Signed-off-by: Chandan Kumar Jha <quic_cjha@quicinc.com>