msm: camera: isp: Include the RD in the SFE top IN resource
In V690, the PIX and RDI paths have been removed from the
SFE top resource, and the RD path has been included to
enable SFE Fetch. The SFE now has only a READ engine for
this target. The hardware acquisition implementation is
based on the TOP input resource, so the RD path has been
added as a top input resource to maintain driver
functionality and achieve SFE offline processing.

This change includes the RD path to enable the
SFE offline engine.

CRs-Fixed: 3987317
Change-Id: I8051209c7a3e2f197af07706b61b8baa73aff3f7
Signed-off-by: Chandan Kumar Jha <quic_cjha@quicinc.com>