msm: camera: isp: Acquire ife lite from same power domain
In Seraph target, two ife lite hw are grouped based on the
power domain. As part of this hardware grouping feature, SW
also need to acquire ife lite from the same power domain
for the same sensor usage group. This commit adds the support
in the ISP driver to acquire the ife lites hw based on the
power domain and sensor usage information.

CRs-Fixed: 4074513
Change-Id: I41263aaaaf76208f4f2fbcb84d5e347e3a51c884
Signed-off-by: Ayush Kumar <quic_ayushkr@quicinc.com>