msm: camera: isp: Add feature support based on kernel version
Lnx.5.4 code is used for both 5.10 and 6.1 MSM kernels.
As result some of the features are not supported on 5.10 kernel.
mem_buf_dma_buf_get_memparcel_hdl function is only available in 6.1
kernel, and is part of CONFIG_SECURE_CAMERA_25
Create a function in cam_compat to load mem_buf_dma_buf_get_memparcel_hdl
only if kernel is CONFIG_SECURE_CAMERA_25 is defined.

CRs-Fixed: 4030420
Change-Id: I89a7ed3762573935ef890d1692e996d0a7ef3eb8
Signed-off-by: Petar Ivanov <quic_pivanov@quicinc.com>