msm: camera: isp: Fix the reconfiguration flag issue for IFE WMs
WMs should be reconfigured for each frame as its dimensions
can be different. This change has fix of the reconfiguration
flag issue which helps when offline isp has different
dimension.
We added the compilation of multiple camera changes to
be invoked by do_compile. Previously, this logic was
moved to do_install, which did not help when there
were compilation errors in the code.

CRs-Fixed: 4093528
Change-Id: I770c6e105f1b02de27f5bf51544b43e5a2c50bb1
Signed-off-by: Chandan Kumar Jha <quic_cjha@quicinc.com>