msm: camera: isp: Update the packer fmt for the FULL output port
In V690 hardware, the input data is in a 12-bit format, meaning a 12-bit
pipeline, so the packet format should be 0x1 according to the hardware
manual. We used a 10-bit pipeline for the mobile target, which is why we
are using 0x3, i.e., Plain8 with LSB to MSB shift for the 10-bit
pipeline.
In V690, the full output port is 3-plane, so we are creating a
2-plane format based on N12 and N21, which is also a 2-plane buffer
format.

CRs-Fixed: 4131305
Change-Id: I60e741cbcf5f4df94579f6a25dfce5a844c6b059
Signed-off-by: Chandan Kumar Jha <quic_cjha@quicinc.com>