msm: camera: common: Fix slab-out-of-bounds in cam packet
Fix slab-out-of-bounds during packet validation. This issue was due to KMD copying the packet from UMD but the length of packet we got was from UMD which could have been corrupted, and then we were reading this copied packet but the lengths here were based on buffer length, due to which out of bound might have occurred when copied length was very small value compared to cam_packet. This change ensures that the packet length value we get from UMD is greater than or equal to size of cam_packet struct.

Change-Id: I917752aae7ba5eb0122ed2fc00b303c9357f5f0c
CRs-Fixed: 3953618
Signed-off-by: Ashish Mishra <quic_ashmish@quicinc.com>