msm: camera: ife: Restrict offline Ife HW acquire to a single instance
In the case of MCX with two VC-DT streams, one of the offline sessions does not have a stat port. However, the fundamental design of offline IFE does not permit proceeding with an uninformed acquire port across the session. It requires that the number of acquire WM ports be consistent across all offline sessions. Therefore, this fix will restrict allocation to only IFE to ensure uniformity across all offline sessions, considering bandwidth and clock constraints. This serves as a workaround to proceed further.

CRs-Fixed: 4157147
Change-Id: I1cd36378e66b2b95070f0001512068a1f39ff1e0
Signed-off-by: Petar Nedev <quic_pnedev@quicinc.com>
Signed-off-by: Dimitar Vachev <quic_dvachev@quicinc.com>