msm: camera: sync: Handle synx import for reusable CSL fences
Currently, when a reusable CSL fence handle is obtained
during a synx import call, the synx object is not
associated with it. Instead, a new CSL handle is created
and associated with the synx handle.

This commit adds a check in synx import to support
associating synx handle with a reusable CSL handle.

CRs-Fixed: 4165619
Change-Id: I29f52a4ccc8a9b430c8293090927e70701c2d0e6
Signed-off-by: Stuti Saxena <stutsaxe@qti.qualcomm.com>