msm: camera: ife: Fix Offline Packet Stall via Pre-Fence Enqueue
In a few rare scenarios, the offline packet submission is delayed. Due
to this accumulation of delays, the final packet may not be submitted
before the corresponding lite node signals buffer-done, which serves as
the input fence for the offline packet. As a result, the packet gets
stuck in the KMD.

This fix addresses the issue by enqueuing the request before registering
the input fence callback, ensuring the offline packet is marked ready
when the callback is triggered.

CRs-Fixed: 4211011
Change-Id: If824e1c640a34632ad15b0fc3737c578a45f7b38
Signed-off-by: Chandan Kumar Jha <cjha@qti.qualcomm.com>