msm: camera: cci: children will not be taken into account while master is suspending
We are seeing issue while switching the cameras from wide to ultra-wide
and vice versa if cameras are getting switch very fast then before
completing power-off sequence of previous camera session(sensor->cci)
another camera is getting turned on and start the power-on sequence
Due to this race condition, CCI genpd turned off is failure as
before turning of gdsc off CCI, another sensor is getting turned on.
This change will ignore the child state and turned off the master gdsc.
As we are managing the state machine of CCI, so when another sensor is
in power on path, it will turn on the CCI gdsc.

CRs-Fixed: 4223266
Change-Id: Ie785fe4941179d6ace4e117b2280d2867acbb8fc
Signed-off-by: Soumen Ghosh <soumeng@qti.qualcomm.com>
(cherry picked from commit 4792ca6539d6204ff6a1dd010babd67a10b859c7)