msm: camera: sensor: Handle kernel compatibility for hrtimer API
The hrtimer_init API has been deprecated and replaced with
hrtimer_setup in the latest kernel version.
This commit updates all relevant references to align with the
new kernel API. This change ensures compatibility with the
updated kernel and avoids build or runtime issues due to
deprecated function usage.

CRs-Fixed: 4225286
Change-Id: I5ad39696078086c993fb7b28243282bca96f78fc
Signed-off-by: Xhoendi Collaku <quic_xcollaku@quicinc.com>
Signed-off-by: Soumen Ghosh <soumeng@qti.qualcomm.com>
Signed-off-by: Joshua Florez <jflorez@qti.qualcomm.com>
Signed-off-by: Lokesh Kumar Aakulu <lkumar@qti.qualcomm.com>
Signed-off-by: Alok Chauhan <alokc@qti.qualcomm.com>