msm: camera: common: Fix slab-out-of-bounds in cam_packet validate
Fix slab-out-of-bounds in cam_packet_util_validate. This issue was due to
KMD copying the packet from UMD but the length of packet we got was from
UMD which could have been corrupted, and then we were reading this copied
packet but the lengths here were based on buffer length, due to which a
out of bound might have occurred when copied length was very small value
compared to cam_packet. This change ensures that the packet length value
we get from UMD is  greater than or equal to size of cam_packet struct.

CRs-Fixed: 4233494
Change-Id: I3532ef0c62753ec3adcd3e6f309a30b7732540ae
Signed-off-by: Pranav Sanwal <psanwal@qti.qualcomm.com>