msm: camera: sensor: lt6911uxc chip bring up
lt6911uxc as a Hdmi-in sensor, this patch supports feature as follows
- Power-on initialization
- lt6911uxc Register management
- Firmware upgrade - Read resolution and fps information
Implementation: lt6911gxc and lt6911uxc belong to the same IC series,
sharing the same driver logic with only register differences.
Based on the original hdmi_bridge driver, the register arrays for each IC are separated,
and the corresponding array is loaded according to the sensor-id.

Change-Id: I8535927c5e8528276bfeaa6c924a22ea8d37ee13
Signed-off-by: Yi Zhang <zhanyi@qti.qualcomm.com>