msm: camera: ife: Add support for RDI crop in dual IFE
This change adds support for RDI crop in dual IFE usecases. The
crop window for RDI is obtained by combining left and right IFE
crop windows and width is calculated based on the crop window.
If crop or drop is needed on RDI, decode format should not be
programmed as payload only in order to enable RPP subpath.

Change-Id: I79dba9f46d7217d85f0d6b78f4e356028e988e58
Signed-off-by: Dongpeng <dongpeng@qti.qualcomm.com>