msm: camera: icp: Fix IPE latency issue during 4K 60fps EIS
On 4K 60fps EIS and multi IPE instances used for different pipeline,
and combined with current clock is same as sum of base clocks, there
is an IPE latency issue and frame drop which related to IPE clk.
If current clock is same as sum of base clocks, it's not overclked.

CRs-Fixed: 4319991
Change-Id: I531bb442cebfd4b0f1be21b5cfade07e09308daf
Signed-off-by: chengxue <chengxue@qti.qualcomm.com>
(cherry picked from commit 814495c4cfeaa9424244caf0109f4b8810cedfb5)