msm: camera: ife: rework camx hw_id to ife_hw conversion
The existing convertion function between IFE hw_id  and internal
numbering will not work correctly if the ife hw_ids are not
contiguous. This can happen if some of IFE/IFE lites are disabled or
carved out from dts.
This change changes the internal numbering to be always contiguous
irrespective of which hw_ids are available.

Change-Id: I802bb44c4493fae98aa9d8ce0dbdca3bbc38b562
Signed-off-by: Petar Sivenov <c_psiven@qti.qualcomm.com>