msm: camera: utils: defined cam_soc_gpio for clarity and consistency
The current code uses struct gpio directly in the Camera SoC GPIO
handling logic. This approach is problematic with newer Linux kernels
because struct gpio is no longer exposed.
We are replacing the usage of struct gpio with struct cam_soc_gpio,
which is an internal abstraction used within the Camera SoC driver.
This ensures that the driver does not rely on kernel-internal
structures that are subject to change.

CRs-Fixed: 4350568
Change-Id: Id12fe34de5d5d55a0691a01975d81f0abc62924c
Signed-off-by: Xhoendi Collaku <xcollaku@qti.qualcomm.com>
Signed-off-by: Soumen Ghosh <soumeng@qti.qualcomm.com>
Signed-off-by: Alok Chauhan <alokc@qti.qualcomm.com>