msm: camera: common: Signal all the buf done for cancelled req
In a flush scenario, the FW pending reqs are returned immediately,
so the FW pending req will be returned earlier than processing req,
which causes the driver to receive out-of-order buffer completion
events. But the driver cannot drop these events because user space
needs them to clean up pending requests.
This change iterates through all the active req and finds the one
that matches the cancelled buffer completion event, so that the buf
done event won't be dropped.

CRs-Fixed: 4351089
Change-Id: I940e2fc55da3cbb4e119b6db8ba7f3fd95ff889c
Signed-off-by: Depeng Shao <depengs@qti.qualcomm.com>