msm: camera: reqmgr: Fix NULL pointer issue in CRM driver
During shutdown the link state and link resources are about to
be removed and destroyed. However, if the event from IFE is
arrived to CRM, it may after aquiring the lock proceed for the
same. During the event trigger processing it may use the resources
which is already freed or in invalid state. To avoid this a link
state check is added, so that when the link is getting destroyed,
no operations are performed further.

CRs-Fixed: 4296784
Change-Id: Idc8e4456b2836267982228a9c483627802dbf5a8
Signed-off-by: Abhilash Kumar <krabhi@qti.qualcomm.com>