msm: camera: reqmgr: Change sequence of locks to fix deadlock
While unlinking devices, the link information is destroyed while
the request lock is held. This ensures that the CRM notification
process, which is enqueued, executes correctly and avoids race
conditions. However, in the current implementation, the session
lock is also held. This can create a deadlock scenario between the
link info destruction and the notification processing, as the
locking sequence during unlinking is opposite to the locks taken
in process_trigger.

This change moves the link information destruction and the request
lock to occur after the session unlock to ensure that the deadlock
scenario is avoided.

CRs-Fixed: 4378607
Change-Id: I0cb502b023012d31ddf0d5a380c387a2e6a9cd64
Signed-off-by: Abhilash Kumar <krabhi@qti.qualcomm.com>